Add "Balances" command and blacklist functionality
A new command, "Balances," was introduced to handle transactions and record the balances. The ability to add users to a blacklist was added to prevent certain individuals from influencing the economy. Minor adjustments were made to other commands. The current version of the project was incremented to represent the new changes made.