Add various features and improvements to Turtle Economy plugin
This update broadens the functionality of the Turtle Economy plugin. This includes implementing the transfer of currency between players and addition of reload and log commands. The update also includes the introduction of an error message for failed balance-change actions, table naming changes in the initial SQL file, and improvements for handling decimals in balance transfers.